Clean up checks for if the source generators should be run on a class
Extra utility methods were added here so simplify checking if a class has a given attribute on it and if any class members have a given attribute. This drastically reduced how much code we needed to have duplicated between source generators to check if it needs to run on a class.